+++ tests/models_test.py
@@ -707,7 +707,7 @@ class PlaylistTest(unittest.TestCase):
         self.assertEqual(playlist.length, 3)
 
     def test_last_modified(self):
-        last_modified = datetime.datetime.now()
+        last_modified = datetime.datetime.utcnow()
         playlist = Playlist(last_modified=last_modified)
         self.assertEqual(playlist.last_modified, last_modified)
         self.assertRaises(
@@ -715,7 +715,7 @@ class PlaylistTest(unittest.TestCase):
 
     def test_with_new_uri(self):
         tracks = [Track()]
-        last_modified = datetime.datetime.now()
+        last_modified = datetime.datetime.utcnow()
         playlist = Playlist(
             uri='an uri', name='a name', tracks=tracks,
             last_modified=last_modified)
@@ -727,7 +727,7 @@ class PlaylistTest(unittest.TestCase):
 
     def test_with_new_name(self):
         tracks = [Track()]
-        last_modified = datetime.datetime.now()
+        last_modified = datetime.datetime.utcnow()
         playlist = Playlist(
             uri='an uri', name='a name', tracks=tracks,
             last_modified=last_modified)
@@ -739,7 +739,7 @@ class PlaylistTest(unittest.TestCase):
 
     def test_with_new_tracks(self):
         tracks = [Track()]
-        last_modified = datetime.datetime.now()
+        last_modified = datetime.datetime.utcnow()
         playlist = Playlist(
             uri='an uri', name='a name', tracks=tracks,
             last_modified=last_modified)
@@ -752,7 +752,7 @@ class PlaylistTest(unittest.TestCase):
 
     def test_with_new_last_modified(self):
         tracks = [Track()]
-        last_modified = datetime.datetime.now()
+        last_modified = datetime.datetime.utcnow()
         new_last_modified = last_modified + datetime.timedelta(1)
         playlist = Playlist(
             uri='an uri', name='a name', tracks=tracks,